refactor(ui): share resolved theme hook across topology and mermaid
Agent-Logs-Url: https://github.com/Mesh-LLM/mesh-llm/sessions/4713c7b8-b96e-48db-8a76-99040e69f1c2
